==> 1_0_preparation.ipynb <==
## Глава 1. Подготовка

Для начала нужно предобработать датасет

import pandas as pd
import numpy as np
import re


def is_phrase(string):
    '''Функция проверяет является ли строка фразой (есть хотя бы 1 пробел)'''
    return " " in string


def is_word(string):
    '''Функция проверяет является ли строка - русскоязычным словом'''
    pattern = re.compile("[ЁёА-я]+")
    return bool(pattern.fullmatch(string))

Надо распарсить файл такого вида:

```
  id  |        stim        |                              word                               
------+--------------------+-----------------------------------------------------------------
    1 | абсолютный         | -
    1 | абсолютный         | 0
    1 | абсолютный         | 100%
    1 | абсолютный         | 100%-ный
    1 | абсолютный         | max
```

Поэтому просто будем делить по `"|"`, и обрезать пустые пробелы.

data = []
with open("./data/to_tag.txt", "r") as in_f:
    for line in in_f.readlines():
        try:
            parts = line.split("|")
            a = parts[0]
            b = parts[1]
            c = "|".join(parts[2:])
            data.append(list(map(str.strip, [a, b, c])))
        except IndexError:
            pass

data = pd.DataFrame(data[1:], columns=data[0])
print(data.shape)
data.head()

Добавим избыточную колонку `"phrase"`, которая является просто соединеным стимулом и реакцией через пробел.

И добавим колонку `auto_tag` - которая будет хранить спецобозначения фраз и НЕ русских слов, чтобы в дальнейшем их не пихать в модели

data["phrase"] = data.stim + " " + data.word
data["auto_tag"] = None
data

Отметим в `auto_tag` фразы и некириллические слова.  

data.loc[data.word.apply(is_phrase), ["auto_tag"]] = "PHRASE"
data.loc[data.word.apply(is_phrase)]

data.loc[(~data.word.apply(is_word)) & (~data.word.apply(is_phrase)), ["auto_tag"]] = 'NULL'
data.loc[(~data.word.apply(is_word)) & (~data.word.apply(is_phrase))]

Работаем с этим.

data[data.word.apply(is_word)].phrase

data.head(15)

Сохраним датасет для будущих поколений (следующих этапов).

data.to_csv('./data/prepared_to_tag.csv', index=False)
