==> 3_0_deep_pavlov.ipynb <==
! pip install deeppavlov

! python -m deeppavlov install morpho_ru_syntagrus_pymorphy

!python -m deeppavlov install squad_bert

import pandas as pd
import numpy as np

import re
import os

os.environ["KERAS_BACKEND"] = "tensorflow"

def is_phrase(string):
    return " " in string

def is_word(string):
    pattern = re.compile("[ЁёА-я]+")
    return bool(pattern.fullmatch(string))

from google.colab import drive
drive.mount('/content/drive')

DATA_FILE = '/content/drive/My Drive/CSC/практика_весна22/data/prepared_to_tag.csv'
SAVE_TO = '/content/drive/My Drive/CSC/практика_весна22/data/deepPavlov.csv'
tagged_cols = ["DP_stim_tag", "DP_word_tag"]

## Глава 3. NN

`deepPavlov`

data = pd.read_csv(DATA_FILE, keep_default_na = False, na_values = [''])
data.head(15)

Создадим колонки для определения частей речи deepPavlov'ым

for c in tagged_cols:
    data[c] = None
data[tagged_cols[1]] = data.auto_tag
data = data.drop(['auto_tag'], axis = 1)
data.head(15)

### DeepPavlov works

from deeppavlov import build_model, configs
model = build_model(configs.morpho_tagger.BERT.morpho_ru_syntagrus_bert, download=True)

def deeppavlov_tag(phrase):
  return ' '.join(model([phrase])[0].split('\t')[3::9])

%%time
fun = deeppavlov_tag
dp_tags = data[data.word.apply(is_word)].phrase.apply(fun)
data.loc[data.word.apply(is_word), tagged_cols[0]] = dp_tags.apply(lambda x: x.split(' ')[0])
data.loc[data.word.apply(is_word), tagged_cols[1]] = dp_tags.apply(lambda x: x.split(' ')[1])
data.loc[~data.word.apply(is_word), tagged_cols[0]] = data.loc[~data.word.apply(is_word), "stim"].apply(fun)
data[:20]

data.to_csv(SAVE_TO, index=False)
