==> 4_0_postprocessing.ipynb <==
import pandas as pd
import numpy as np
from tqdm import tqdm

pd.options.mode.chained_assignment = None

def flatten(t):
    return [item for sublist in t for item in sublist]

DATA_FILES = [
    "./data/natasha_pymorphy.csv",
    "./data/rnnmorph.csv",
    "./data/deepPavlov.csv",
]
TEST_FILE = './data/test.txt'
COLS = [
    ["morph_stim_tag", "morph_word_tag", "nata_stim_tag", "nata_word_tag"],
    ["rnn_stim_tag", "rnn_word_tag"],
    ["DP_stim_tag", "DP_word_tag"],
]

datas = []
for file in DATA_FILES:
    datas.append(pd.read_csv(file))

merged_data = pd.concat([datas[0], datas[1][COLS[1]], datas[2][COLS[2]]], axis = 1)
merged_data.head()

# merged_data[merged_data[COLS[2][0]]=="X"]

#https://universaldependencies.org/u/pos/all.html#al-u-pos/PART

tag2tag = {
    "A": "A",
    "ADJ": "A",
    "ADP": "nan", # adposition 2,...
    "ADV": "ADV",
    "AUX": "nan", # auxilary
    "CCONJ": "CONJ",
    "COMP": "COMP", #                                         МОЖНО?
    "CONJ": "CONJ",
    "DET": "nan", # determiner
    "H": "nan", # ?????
    "INTJ": "INTJ",
    "NOUN": "S",
    "NUM": "NUM",
    "PART": "PART",
    "PHRASE": "PHRASE",
    "PR": "PR",
    "PRAEDIC": "PR",
    "PRON": "PRO",
    "PROPN": "S",
    "PUNCT": "nan", #PUNCT: punctuation
    "S": "S",
    "SCONJ": "CONJ",
    "SPRO": "PRO",
    "UNKN": "UNKN",
    "V": "V",
    "VERB": "V",
    "X": "nan", #? X: other
    "nan": "nan",
}

stim_tag_cols = flatten(COLS)[0::2]


def split_str(x):
    try:
        return x.split("||")
    except AttributeError:
        return [x]


all_tags = []
for col in stim_tag_cols:
    all_tags.append(np.unique(flatten(list(merged_data[col].apply(split_str)))))
all_tags_u = np.unique(flatten(all_tags))
# for tag in all_tags:
#     print(f'\'{tag}\':\'\',')
all_tags_u

test = pd.read_csv(TEST_FILE, header=None)
test.columns = ['test']
test['word'] = test.test.apply(lambda x: x.split('_')[0])
test['tag'] = test.test.apply(lambda x: x.split('_')[1])
test

checked_dataset = merged_data[['stim']+stim_tag_cols]
checked_dataset

checked_cols = []

for col in stim_tag_cols:
    checked_cols.append(f'{col}_check')
    checked_dataset.loc[:,[checked_cols[-1]]] = None
checked_dataset

def make_matrix():
    t = np.unique(list(tag2tag.values()))
    n = len(t)
    matrix = pd.DataFrame(np.zeros((n, n)), index=t, columns=t, dtype=int)
    return matrix

def check(pred):
    def map_tag2tag(x):
        try:
            return tag2tag[x]
        except KeyError:
            return "nan"

    predicted_tags = set(list(map(map_tag2tag, pred.split("||"))))
    true_tags = set(tags)
    for t in true_tags:
        for p in predicted_tags:
            matrix.loc[t][p] += 1
    return len(predicted_tags & true_tags) > 0


matrixes = dict()
for col in stim_tag_cols:
    matrixes.update({col: make_matrix()})
for key, val in tqdm(test.iterrows()):
    word = val.word
    tag = val.tag
    tags = tag.split("||")
    trues = dict()
    for col in stim_tag_cols:
        matrix = matrixes[col]
        row_flag = checked_dataset.stim == word
        checked_dataset.loc[row_flag, [f"{col}_check"]] = checked_dataset[row_flag][
            col
        ].apply(check)

checked_dataset

check_data = checked_dataset[~checked_dataset.DP_stim_tag_check.isna()]
n = check_data.shape[0]
np.sum(check_data[checked_cols], axis = 0)/n

for col in stim_tag_cols:
    print(col)
    m = matrixes[col]
#     m[(m!= 0).any(axis=1),(m!= 0).any(axis=0)]
    display(m.loc[(m!= 0).any(axis=1),(m!= 0).any(axis=0)])
